fix(vterm): correct schar_T size mismatch causing terminal-open busy-loop hang
## Root cause

`SChar` in the Rust vterm mirror was `u64` (8 bytes) but the C type it
mirrors — `schar_T` in `types_defs.h` — is `typedef uint32_t schar_T`
(4 bytes).

This made the Rust `VTermScreenCell` 28 bytes while the C struct is 24
bytes. Specifically, the `width` field landed at offset 8 in Rust but
at offset 4 in C. When `rs_fetch_row` called `vterm_screen_get_cell`
(C), which wrote `width = 1` at offset 4, the Rust code read
`cell.width` at offset 8 — hitting alignment padding — and got 0.

Since `col += c_int::from(cell.width)` added 0 every iteration, `col`
never advanced past 0, creating a 99.8%-CPU infinite loop every time
`:terminal` was opened.

## Fix

- Change `pub type SChar = u64` → `pub type SChar = u32` in
  `src/nvim-rs/vterm/src/lib.rs` (all three vterm structs with a
  `schar` field — `VTermScreenCell`, `VTermGlyphInfo`, `ScreenCell` —
  now have the correct 4-byte `schar_T`-equivalent field).
- Fix the continuation-cell sentinel in `screen.rs`: was `u64::MAX`,
  must be `u32::MAX` to match the C sentinel `(uint32_t)-1` written by
  `screen.c`.
- Add a Rust compile-time guard:
  `assert!(size_of::<VTermScreenCell>() == 24)` in `lib.rs`.
- Add a matching C guard:
  `_Static_assert(sizeof(VTermScreenCell) == 24, ...)` in
  `vterm_defs.h`.

Verified: terminal functional tests now complete without hanging
(previously timed out at 90 s with 99.8% CPU spin); all 4339 Rust unit
tests pass; `just check` (build + clippy + fmt + tests) green.

+++ src/nvim-rs/vterm/src/screen.rs
@@ -313,16 +313,19 @@ impl ScreenCell {
     }
 
     /// Check if this is a continuation cell (part of a wide character)
+    ///
+    /// Matches C sentinel `(uint32_t)-1` used in screen.c.
     #[inline]
-    #[allow(clippy::cast_possible_wrap)]
     pub const fn is_continuation(&self) -> bool {
-        self.schar == u64::MAX
+        self.schar == u32::MAX
     }
 
     /// Mark this cell as a continuation cell
+    ///
+    /// Matches C sentinel `(uint32_t)-1` used in screen.c.
     #[inline]
     pub fn mark_continuation(&mut self) {
-        self.schar = u64::MAX;
+        self.schar = u32::MAX;
     }
 }
